Modularized code, implemented runner
